test: Add IPv6 ClusterIP test case to validate bpf_lxc
Previous commit dropped the manual DualStack test cases which we run on
the 4.9 CI job. The job runs with KPR=disabled, to it was testing IPv6
ClusterIP for the per-packet LB (bpf_lxc).

The ClusterIP IPv6 in bpf_lxc should be covered by the smoke-v6 suite
run in the GH action. However, it's unclear whether the suite is running
with KPR=disabled or enabled (it's up to cilium-cli). So to make sure
that we don't accidentally loose the coverage, add the IPv6 ClusterIP
tests to the test case which disables the socket-lb for pod netns.

Signed-off-by: Martynas Pumputis <m@lambda.lt>