test: Remove disabled socket-lb same node case
This commit removes the "Checks ClusterIP connectivity on the same node"
test case which was wrongly assuming that it queried the service
endpoint running on the same node.

Instead of fixing the test case, the commit removes it, as testing on
the same node would not give any new coverage wrt existing "across the
nodes" test case.

Signed-off-by: Martynas Pumputis <m@lambda.lt>