ipcache: Rework metadata cache for any labels
Rework the internals of the "IDMD" Identity <-> Metadata cache so that
it can store multiple pieces of information coming from different
sources, rather than only labels coming from the kube-apiserver identity
feature. Initially in this commit, this only includes one feature source
which only stores labels internally, which is what's necessary for the
current user of this code.

In future, we anticipate extending the core info structure here to
include not just labels, but also auxiliary data like port name mappings
and encryption indexes shared by other nodes.

Information is keyed by the UID, typically a UID of a resource from
kubernetes, which allows tracing back the source of the information, as
well as convenient insertion/deletion by UID. This can mean that the
same information may be stored multiple times by different UID inside
the structure, however this is not expected to be the normal case, so
the memory cost for this is expected to be minimal. On the other hand,
directly associating UIDs should allow easier tracing of the source of
the information for debugging & tracing purposes.

Co-authored-by: Chris Tarazi <chris@isovalent.com>
Signed-off-by: Joe Stringer <joe@cilium.io>