datapath: Bring back WG pod2pod behavior when tunnel != disabled
Before changing the WG integration's behavior, when running in the
tunneling mode, a pod2pod@remote-node traffic escaped the bpf_overlay's
tunneling, and was encapsulated once by the WG tunnel netdev.

To be compatible with this < v1.12 behavior, this commit adds the
redirect to the WG tunnel to the __encap_and_redirect_with_nodeid()
function which is eventually called in the pod2pod packet path.

Signed-off-by: Martynas Pumputis <m@lambda.lt>