bpf: egressgw: move bpf_lxc logic into bpf_host
This commit moves the egress gateway logic to encapsulate and forward
traffic to a gateway node from bpf_lxc/handle_ipv4_from_lxc() to
bpf_host/to_netdev().

This fixes the incompatibility between egress gateway and L7 policies,
as currently, when support for L7 policies is enabled, traffic would be
redirected to the proxy before the logic in bpf_lxc can inspect it to
determine if it should be forwarded, resulting in no traffic being
forwarded to the egress gateway node.

While at it, also skip redirecting traffic over the tunnel if the local
node is supposed to act as egress gateway for the traffic (as otherwise
packets would end up in a network loop).

Signed-off-by: Gilberto Bertin <jibi@cilium.io>